sua lai giao dien setting, fix bug khong cap nhap tong income/expense

+++ 04_Implement/02_Source/app/src/main/java/com/example/thang/smartmoney/database/Database.java
@@ -109,6 +109,12 @@ public class Database extends SQLiteOpenHelper {
         {
             Log.d("database", "Upgrade from version " + oldVersion + " to " + newVersion);
 
+            if (newVersion == 4) {
+                // version 0.4, no need delete old data
+                InitDB(db, R.raw.initdb);
+                return;
+            }
+
             try {
                 db.execSQL("DROP TABLE IF EXISTS " + TABLE.GiaoDich);
                 db.execSQL("DROP TABLE IF EXISTS " + TABLE.Category);
